// src/app/models/employee.model.ts
export class Employee{

    id: number | any;
    name: string | any;
    gender: string | any;
    contactPreference: string | any;
    email?: string | any;
    dateOfBirth: Date | any;
    department: string | any;
    phoneNumber?: number| any;;
    isActive: boolean | any;
    photoPath?: string | any;
    password: string | any;
    confirmPassword: string | any;
}